refactor / create snapshot and manager classes
