+++ subgraphs/compound-v3/protocols/compound-v3/config/templates/compound.v3.template.yaml
@@ -98,6 +98,9 @@ templates:
         - event: AbsorbCollateral(indexed address,indexed address,indexed address,uint256,uint256)
           handler: handleAbsorbCollateral
           receipt: true
+        - event: Transfer(indexed address,indexed address,uint256)
+          handler: handleTransfer
+          receipt: true
       # callHandlers:
       #   - function: transferBase(indexed address,indexed address,uint256)
       #     handler: handleTransferBase
