+++ subgraphs/compound-v3/src/sdk/snapshots.ts
@@ -1,230 +1,66 @@
 import {
-  Address,
-  Bytes,
-  BigInt,
-  ethereum,
   BigDecimal,
+  BigInt,
+  Bytes,
+  ethereum,
   log,
 } from "@graphprotocol/graph-ts";
 import {
   ActiveAccount,
-  Borrow,
-  Deposit,
   Fee,
   FinancialsDailySnapshot,
-  Flashloan,
   InterestRate,
   LendingProtocol,
-  Liquidate,
   Market,
   MarketDailySnapshot,
   MarketHourlySnapshot,
-  Oracle,
-  Repay,
   RevenueDetails,
-  Token,
-  Transfer,
   UsageMetricsDailySnapshot,
   UsageMetricsHourlySnapshot,
-  Withdraw,
 } from "../../generated/schema";
-import { Versions } from "../versions";
-import { AccountClass } from "./account";
 import {
   AccountActivity,
   BIGDECIMAL_ZERO,
   BIGINT_ZERO,
-  exponentToBigDecimal,
   INT_ONE,
   INT_ZERO,
-  PositionSide,
-  ProtocolType,
-  RevenueSource,
   SECONDS_PER_DAY,
   SECONDS_PER_HOUR,
   TransactionType,
 } from "./constants";
-import { TokenClass } from "./token";
 
-/**
- * This file contains the MarketClass, which is used to
- * make all of the storage changes that occur in a given market.
- *
- * You can think of this as an abstraction so the developer doesn't
- * need to think about all of the detailed storage changes that occur.
- *
- * Schema Version: 3.0.0
- * Last Updated: Dec 4, 2022
- * Author(s):
- *  - @dmelotik
- */
-export class MarketClass {
-  private event!: ethereum.Event;
-
-  // entities
-  public protocol!: LendingProtocol;
-  private market!: Market;
-  private inputToken!: TokenClass;
-  private account!: AccountClass;
-  private oracle!: Oracle;
-
-  // snapshots
+export class SnapshotManager {
   private marketHourlySnapshot!: MarketHourlySnapshot;
   private marketDailySnapshot!: MarketDailySnapshot;
   private financialSnapshot!: FinancialsDailySnapshot;
   private usageHourlySnapshot!: UsageMetricsHourlySnapshot;
   private usageDailySnapshot!: UsageMetricsDailySnapshot;
 
-  // instantiate MarketClass (create new Market if necessary)
+  private event: ethereum.Event;
+  private market: Market;
+  private protocol: LendingProtocol;
+
   constructor(
-    marketID: Bytes,
-    inputToken: Bytes,
     event: ethereum.Event,
-    protocolData: ProtocolData
+    protocol: LendingProtocol,
+    market: Market
   ) {
-    this.protocol = this.getOrCreateLendingProtocol(protocolData);
-    this.inputToken = new TokenClass(inputToken, event);
-    let _market = Market.load(marketID);
-
-    // create new market
-    if (!_market) {
-      _market = new Market(marketID);
-      _market.protocol = this.protocol.id;
-      _market.isActive = true;
-      _market.canBorrowFrom = false; // default
-      _market.canUseAsCollateral = false; // default
-      _market.maximumLTV = BIGDECIMAL_ZERO; // default
-      _market.liquidationThreshold = BIGDECIMAL_ZERO; // default
-      _market.liquidationPenalty = BIGDECIMAL_ZERO; // default
-      _market.canIsolate = false; // default
-      _market.inputToken = this.inputToken.getToken().id;
-      _market.inputTokenBalance = BIGINT_ZERO;
-      _market.inputTokenPriceUSD = BIGDECIMAL_ZERO;
-      _market.totalValueLockedUSD = BIGDECIMAL_ZERO;
-      _market.cumulativeSupplySideRevenueUSD = BIGDECIMAL_ZERO;
-      _market.cumulativeProtocolSideRevenueUSD = BIGDECIMAL_ZERO;
-      _market.cumulativeTotalRevenueUSD = BIGDECIMAL_ZERO;
-      _market.totalDepositBalanceUSD = BIGDECIMAL_ZERO;
-      _market.cumulativeDepositUSD = BIGDECIMAL_ZERO;
-      _market.totalBorrowBalanceUSD = BIGDECIMAL_ZERO;
-      _market.cumulativeBorrowUSD = BIGDECIMAL_ZERO;
-      _market.cumulativeLiquidateUSD = BIGDECIMAL_ZERO;
-      _market.cumulativeTransferUSD = BIGDECIMAL_ZERO;
-      _market.cumulativeFlashloanUSD = BIGDECIMAL_ZERO;
-      _market.transactionCount = INT_ZERO;
-      _market.depositCount = INT_ZERO;
-      _market.withdrawalCount = INT_ZERO;
-      _market.borrowCount = INT_ZERO;
-      _market.repayCount = INT_ZERO;
-      _market.liquidationCount = INT_ZERO;
-      _market.transferCount = INT_ZERO;
-      _market.flashloanCount = INT_ZERO;
-
-      _market.cumulativeUniqueUsers = INT_ZERO;
-      _market.cumulativeUniqueDepositors = INT_ZERO;
-      _market.cumulativeUniqueBorrowers = INT_ZERO;
-      _market.cumulativeUniqueLiquidators = INT_ZERO;
-      _market.cumulativeUniqueLiquidatees = INT_ZERO;
-      _market.cumulativeUniqueTransferrers = INT_ZERO;
-      _market.cumulativeUniqueFlashloaners = INT_ZERO;
-
-      _market.createdTimestamp = event.block.timestamp;
-      _market.createdBlockNumber = event.block.number;
-
-      _market.positionCount = INT_ZERO;
-      _market.openPositionCount = INT_ZERO;
-      _market.closedPositionCount = INT_ZERO;
-      _market.lendingPositionCount = INT_ZERO;
-      _market.borrowingPositionCount = INT_ZERO;
-      _market.save();
-
-      // add market to protocol
-      const markets = this.protocol._markets;
-      markets.push(_market.id);
-      this.protocol.markets = markets;
-      this.protocol.totalPoolCount += INT_ONE;
-      this.protocol.save();
-    }
-    this.market = _market;
     this.event = event;
+    this.protocol = protocol;
+    this.market = market;
 
-    // load snapshots
-    this.getOrCreateMarketHourlySnapshot();
-    this.getOrCreateMarketDailySnapshot();
-    this.getOrCreateFinancials();
-    this.getOrCreateUsageDailySnapshot();
-    this.getOrCreateUsageHourlySnapshot();
-
-    // load oracle
-    if (this.market.oracle) {
-      this.oracle = Oracle.load(this.market.oracle!)!;
-    }
+    this.createOrUpdateMarketHourlySnapshot();
+    this.createOrUpdateMarketDailySnapshot();
+    this.createOrUpdateFinancials();
+    this.createOrUpdateUsageDailySnapshot();
+    this.createOrUpdateUsageHourlySnapshot();
   }
 
-  /////////////////
-  //// Getters ////
-  /////////////////
+  ///////////////////
+  ///// Getters /////
+  ///////////////////
 
-  getOrCreateLendingProtocol(data: ProtocolData): LendingProtocol {
-    let protocol = LendingProtocol.load(data.protocolID);
-    if (!protocol) {
-      protocol = new LendingProtocol(data.protocolID);
-      protocol.protocol = data.protocol;
-      protocol.name = data.name;
-      protocol.slug = data.slug;
-      protocol.network = data.network;
-      protocol.type = ProtocolType.LENDING;
-      protocol.lendingType = data.lendingType;
-      protocol.lenderPermissionType = data.lenderPermissionType;
-      protocol.borrowerPermissionType = data.borrowerPermissionType;
-      protocol.riskType = data.riskType;
-      protocol.collateralizationType = data.collateralizationType;
-
-      protocol.cumulativeUniqueUsers = INT_ZERO;
-      protocol.cumulativeUniqueDepositors = INT_ZERO;
-      protocol.cumulativeUniqueBorrowers = INT_ZERO;
-      protocol.cumulativeUniqueLiquidators = INT_ZERO;
-      protocol.cumulativeUniqueLiquidatees = INT_ZERO;
-      protocol.totalValueLockedUSD = BIGDECIMAL_ZERO;
-      protocol.cumulativeSupplySideRevenueUSD = BIGDECIMAL_ZERO;
-      protocol.cumulativeProtocolSideRevenueUSD = BIGDECIMAL_ZERO;
-      protocol.cumulativeTotalRevenueUSD = BIGDECIMAL_ZERO;
-      protocol.totalDepositBalanceUSD = BIGDECIMAL_ZERO;
-      protocol.cumulativeDepositUSD = BIGDECIMAL_ZERO;
-      protocol.totalBorrowBalanceUSD = BIGDECIMAL_ZERO;
-      protocol.cumulativeBorrowUSD = BIGDECIMAL_ZERO;
-      protocol.cumulativeLiquidateUSD = BIGDECIMAL_ZERO;
-      protocol.totalPoolCount = INT_ZERO;
-      protocol.openPositionCount = INT_ZERO;
-      protocol.cumulativePositionCount = INT_ZERO;
-      protocol.transactionCount = INT_ZERO;
-      protocol.depositCount = INT_ZERO;
-      protocol.withdrawalCount = INT_ZERO;
-      protocol.borrowCount = INT_ZERO;
-      protocol.repayCount = INT_ZERO;
-      protocol.liquidationCount = INT_ZERO;
-      protocol.transferCount = INT_ZERO;
-      protocol.flashloanCount = INT_ZERO;
-      protocol._markets = [];
-    }
-
-    protocol.schemaVersion = Versions.getSchemaVersion();
-    protocol.subgraphVersion = Versions.getSubgraphVersion();
-    protocol.methodologyVersion = Versions.getMethodologyVersion();
-    protocol.save();
-
-    return protocol;
-  }
-
-  getMarket(): Market {
-    return this.market;
-  }
-
-  getInputToken(): Token {
-    return this.inputToken.getToken();
-  }
-
-  getOrCreateMarketHourlySnapshot(): void {
+  private createOrUpdateMarketHourlySnapshot(): void {
     const hours = this.event.block.timestamp.toI32() / SECONDS_PER_HOUR;
     const id = this.market.id.concat(Bytes.fromI32(hours));
     let snapshot = MarketHourlySnapshot.load(id);
@@ -283,7 +119,7 @@ export class MarketClass {
     this.marketHourlySnapshot = snapshot;
   }
 
-  getOrCreateMarketDailySnapshot(): void {
+  private createOrUpdateMarketDailySnapshot(): void {
     const days = this.event.block.timestamp.toI32() / SECONDS_PER_DAY;
     const id = this.market.id.concat(Bytes.fromI32(days));
     let snapshot = MarketDailySnapshot.load(id);
@@ -376,7 +212,7 @@ export class MarketClass {
     this.marketDailySnapshot = snapshot;
   }
 
-  getOrCreateFinancials(): void {
+  createOrUpdateFinancials(): void {
     const days = this.event.block.timestamp.toI32() / SECONDS_PER_DAY;
     const id = Bytes.fromI32(days);
     let snapshot = FinancialsDailySnapshot.load(id);
@@ -410,8 +246,8 @@ export class MarketClass {
       this.protocol.cumulativeProtocolSideRevenueUSD;
     snapshot.cumulativeTotalRevenueUSD =
       this.protocol.cumulativeTotalRevenueUSD;
-    snapshot.revenueDetails = this.market.revenueDetails
-      ? this.getSnapshotRevenueDetails(this.market.revenueDetails!, days)
+    snapshot.revenueDetails = this.protocol.revenueDetails
+      ? this.getSnapshotRevenueDetails(this.protocol.revenueDetails!, days)
       : null;
     snapshot.totalDepositBalanceUSD = this.protocol.totalDepositBalanceUSD;
     snapshot.cumulativeDepositUSD = this.protocol.cumulativeDepositUSD;
@@ -423,7 +259,7 @@ export class MarketClass {
     this.financialSnapshot = snapshot;
   }
 
-  getOrCreateUsageDailySnapshot(): void {
+  createOrUpdateUsageDailySnapshot(): void {
     const days = this.event.block.timestamp.toI32() / SECONDS_PER_DAY;
     const id = Bytes.fromI32(days);
     let snapshot = UsageMetricsDailySnapshot.load(id);
@@ -466,7 +302,7 @@ export class MarketClass {
     this.usageDailySnapshot = snapshot;
   }
 
-  getOrCreateUsageHourlySnapshot(): void {
+  createOrUpdateUsageHourlySnapshot(): void {
     const hours = this.event.block.timestamp.toI32() / SECONDS_PER_HOUR;
     const id = Bytes.fromI32(hours);
     let snapshot = UsageMetricsHourlySnapshot.load(id);
@@ -491,486 +327,11 @@ export class MarketClass {
     this.usageHourlySnapshot = snapshot;
   }
 
-  getOrCreateOracle(
-    oracleAddress: Address,
-    isUSD: boolean,
-    source?: string
-  ): Oracle {
-    const oracleID = this.market.id.concat(this.market.inputToken);
-    let oracle = Oracle.load(oracleID);
-    if (!oracle) {
-      oracle = new Oracle(oracleID);
-      oracle.market = this.market.id;
-      oracle.blockCreated = this.event.block.number;
-      oracle.timestampCreated = this.event.block.timestamp;
-      oracle.isActive = true;
-    }
-    oracle.oracleAddress = oracleAddress;
-    oracle.isUSD = isUSD;
-    if (source) {
-      oracle.oracleSource = source;
-    }
-    oracle.save();
-    this.oracle = oracle;
+  ////////////////////
+  ///// Updaters /////
+  ////////////////////
 
-    return oracle;
-  }
-
-  getOracleAddress(): Address {
-    return Address.fromBytes(this.oracle.oracleAddress);
-  }
-
-  getOrUpdateRate(
-    rateSide: string,
-    rateType: string,
-    interestRate: BigDecimal
-  ): InterestRate {
-    const interestRateID = rateSide
-      .concat("-")
-      .concat(rateType)
-      .concat("-")
-      .concat(this.market.id.toHexString());
-    let rate = InterestRate.load(interestRateID);
-    if (!rate) {
-      rate = new InterestRate(interestRateID);
-      rate.side = rateSide;
-      rate.type = rateType;
-    }
-    rate.rate = interestRate;
-    rate.save();
-
-    if (!this.market.rates) {
-      this.market.rates = [];
-    }
-
-    if (this.market.rates!.indexOf(interestRateID) == -1) {
-      this.market.rates!.push(interestRateID);
-    }
-    this.market.save();
-
-    return rate;
-  }
-
-  getAddress(): Address {
-    return Address.fromBytes(this.market.id);
-  }
-
-  getSnapshotRates(rates: string[], timeSuffix: string): string[] {
-    const snapshotRates: string[] = [];
-    for (let i = 0; i < rates.length; i++) {
-      const rate = InterestRate.load(rates[i]);
-      if (!rate) {
-        log.warning("[getSnapshotRates] rate {} not found, should not happen", [
-          rates[i],
-        ]);
-        continue;
-      }
-
-      // create new snapshot rate
-      const snapshotRateId = rates[i].concat("-").concat(timeSuffix);
-      const snapshotRate = new InterestRate(snapshotRateId);
-      snapshotRate.rate = rate.rate;
-      if (rate.maturityBlock) snapshotRate.maturityBlock = rate.maturityBlock;
-      snapshotRate.side = rate.side;
-      snapshotRate.type = rate.type;
-      if (rate.tranche) snapshotRate.tranche = rate.tranche;
-      snapshotRate.save();
-      snapshotRates.push(snapshotRateId);
-    }
-    return snapshotRates;
-  }
-
-  getSnapshotFees(fees: string[], timeSuffix: string): string[] {
-    const snapshotFees: string[] = [];
-    for (let i = 0; i < fees.length; i++) {
-      const rate = InterestRate.load(fees[i]);
-      if (!rate) {
-        log.error("[getSnapshotFees] fee {} not found, should not happen", [
-          fees[i],
-        ]);
-        continue;
-      }
-
-      // create new snapshot rate
-      const snapshotFeeID = fees[i].concat("-").concat(timeSuffix);
-      const snapshotFee = new Fee(snapshotFeeID);
-      snapshotFee.rate = rate.rate;
-      snapshotFee.type = rate.type;
-      snapshotFee.save();
-      snapshotFees.push(snapshotFee.id);
-    }
-    return snapshotFees;
-  }
-
-  getSnapshotRevenueDetails(currID: Bytes, timeSuffix: i32): Bytes | null {
-    const currDetails = RevenueDetails.load(currID);
-    if (!currDetails) {
-      log.error(
-        "[getRevenueDetailsSnapshot] Cannot find revenue details id: {}",
-        [currID.toHexString()]
-      );
-      return null;
-    }
-
-    const newDetails = new RevenueDetails(
-      currDetails.id.concat(Bytes.fromI32(timeSuffix))
-    );
-    newDetails.sources = currDetails.sources;
-    newDetails.amountsUSD = currDetails.amountsUSD;
-    newDetails.save();
-
-    return newDetails.id;
-  }
-
-  getOrCreateRevenueDetails(id: Bytes): RevenueDetails {
-    let details = RevenueDetails.load(id);
-    if (!details) {
-      details = new RevenueDetails(id);
-      details.sources = [
-        RevenueSource.BORROW_INTEREST,
-        RevenueSource.FLASHLOAN_FEE,
-        RevenueSource.LIQUIDATION_FEE,
-      ];
-      details.amountsUSD = [BIGDECIMAL_ZERO, BIGDECIMAL_ZERO, BIGDECIMAL_ZERO];
-      details.save();
-    }
-
-    return details;
-  }
-
-  //////////////////
-  //// Creators ////
-  //////////////////
-
-  createDeposit(
-    asset: Address,
-    account: Address,
-    amount: BigInt,
-    amountUSD: BigDecimal,
-    newBalance: BigInt,
-    interestType: string | null = null
-  ): Deposit {
-    this.account = new AccountClass(
-      account,
-      this.market,
-      this.protocol,
-      this.event
-    );
-    if (this.account.newAccount()) {
-      this.protocol.cumulativeUniqueUsers += INT_ONE;
-      this.protocol.save();
-    }
-    this.account.addPosition(
-      this.market.inputToken,
-      newBalance,
-      PositionSide.COLLATERAL,
-      TransactionType.DEPOSIT,
-      this.market.inputTokenPriceUSD,
-      interestType
-    );
-
-    const deposit = new Deposit(
-      this.event.transaction.hash.concatI32(this.event.logIndex.toI32())
-    );
-    deposit.hash = this.event.transaction.hash;
-    deposit.nonce = this.event.transaction.nonce;
-    deposit.logIndex = this.event.logIndex.toI32();
-    deposit.gasPrice = this.event.transaction.gasPrice;
-    deposit.gasUsed = this.event.receipt ? this.event.receipt!.gasUsed : null;
-    deposit.gasLimit = this.event.transaction.gasLimit;
-    deposit.blockNumber = this.event.block.number;
-    deposit.timestamp = this.event.block.timestamp;
-    deposit.account = account;
-    deposit.market = this.market.id;
-    deposit.position = this.account.getPositionID();
-    deposit.asset = asset;
-    deposit.amount = amount;
-    deposit.amountUSD = amountUSD;
-    deposit.save();
-
-    this.updateTransactionData(TransactionType.DEPOSIT, amount, amountUSD);
-    this.updateSnapshotUsage(TransactionType.DEPOSIT, account);
-
-    return deposit;
-  }
-
-  createWithdraw(
-    asset: Address,
-    account: Address,
-    amount: BigInt,
-    amountUSD: BigDecimal
-  ): Withdraw {
-    this.account = new AccountClass(
-      account,
-      this.market,
-      this.protocol,
-      this.event
-    );
-    if (this.account.newAccount()) {
-      this.protocol.cumulativeUniqueUsers += INT_ONE;
-      this.protocol.save();
-    }
-
-    const withdraw = new Withdraw(
-      this.event.transaction.hash.concatI32(this.event.logIndex.toI32())
-    );
-    withdraw.hash = this.event.transaction.hash;
-    withdraw.nonce = this.event.transaction.nonce;
-    withdraw.logIndex = this.event.logIndex.toI32();
-    withdraw.gasPrice = this.event.transaction.gasPrice;
-    withdraw.gasUsed = this.event.receipt ? this.event.receipt!.gasUsed : null;
-    withdraw.gasLimit = this.event.transaction.gasLimit;
-    withdraw.blockNumber = this.event.block.number;
-    withdraw.timestamp = this.event.block.timestamp;
-    withdraw.account = account;
-    withdraw.market = this.market.id;
-    withdraw.position = account; // TODO add position
-    withdraw.asset = asset;
-    withdraw.amount = amount;
-    withdraw.amountUSD = amountUSD;
-    withdraw.save();
-
-    this.updateTransactionData(TransactionType.WITHDRAW, amount, amountUSD);
-    this.updateSnapshotUsage(TransactionType.WITHDRAW, account);
-
-    return withdraw;
-  }
-
-  createBorrow(
-    asset: Address,
-    account: Address,
-    amount: BigInt,
-    amountUSD: BigDecimal
-  ): Borrow {
-    this.account = new AccountClass(
-      account,
-      this.market,
-      this.protocol,
-      this.event
-    );
-    if (this.account.newAccount()) {
-      this.protocol.cumulativeUniqueUsers += INT_ONE;
-      this.protocol.save();
-    }
-
-    const borrow = new Borrow(
-      this.event.transaction.hash.concatI32(this.event.logIndex.toI32())
-    );
-    borrow.hash = this.event.transaction.hash;
-    borrow.nonce = this.event.transaction.nonce;
-    borrow.logIndex = this.event.logIndex.toI32();
-    borrow.gasPrice = this.event.transaction.gasPrice;
-    borrow.gasUsed = this.event.receipt ? this.event.receipt!.gasUsed : null;
-    borrow.gasLimit = this.event.transaction.gasLimit;
-    borrow.blockNumber = this.event.block.number;
-    borrow.timestamp = this.event.block.timestamp;
-    borrow.account = account;
-    borrow.market = this.market.id;
-    borrow.position = account; // TODO add position
-    borrow.asset = asset;
-    borrow.amount = amount;
-    borrow.amountUSD = amountUSD;
-    borrow.save();
-
-    this.updateTransactionData(TransactionType.BORROW, amount, amountUSD);
-    this.updateSnapshotUsage(TransactionType.BORROW, account);
-
-    return borrow;
-  }
-
-  createRepay(
-    asset: Address,
-    account: Address,
-    amount: BigInt,
-    amountUSD: BigDecimal
-  ): Repay {
-    this.account = new AccountClass(
-      account,
-      this.market,
-      this.protocol,
-      this.event
-    );
-    if (this.account.newAccount()) {
-      this.protocol.cumulativeUniqueUsers += INT_ONE;
-      this.protocol.save();
-    }
-
-    const repay = new Repay(
-      this.event.transaction.hash.concatI32(this.event.logIndex.toI32())
-    );
-    repay.hash = this.event.transaction.hash;
-    repay.nonce = this.event.transaction.nonce;
-    repay.logIndex = this.event.logIndex.toI32();
-    repay.gasPrice = this.event.transaction.gasPrice;
-    repay.gasUsed = this.event.receipt ? this.event.receipt!.gasUsed : null;
-    repay.gasLimit = this.event.transaction.gasLimit;
-    repay.blockNumber = this.event.block.number;
-    repay.timestamp = this.event.block.timestamp;
-    repay.account = account;
-    repay.market = this.market.id;
-    repay.position = account; // TODO add position
-    repay.asset = asset;
-    repay.amount = amount;
-    repay.amountUSD = amountUSD;
-    repay.save();
-
-    this.updateTransactionData(TransactionType.REPAY, amount, amountUSD);
-    this.updateSnapshotUsage(TransactionType.REPAY, account);
-
-    return repay;
-  }
-
-  createLiquidate(
-    asset: Address,
-    liquidator: Address,
-    liquidatee: Address,
-    amount: BigInt,
-    amountUSD: BigDecimal,
-    profitUSD: BigDecimal
-  ): Liquidate {
-    this.account = new AccountClass(
-      liquidator,
-      this.market,
-      this.protocol,
-      this.event
-    );
-    if (this.account.newAccount()) {
-      this.protocol.cumulativeUniqueUsers += INT_ONE;
-      this.protocol.save();
-    }
-    const liquidateeAccount = new AccountClass(
-      liquidatee,
-      this.market,
-      this.protocol,
-      this.event
-    );
-    // liquidatees are not considered users since they are not spending gas for the transaction
-
-    const liquidate = new Liquidate(
-      this.event.transaction.hash.concatI32(this.event.logIndex.toI32())
-    );
-    liquidate.hash = this.event.transaction.hash;
-    liquidate.nonce = this.event.transaction.nonce;
-    liquidate.logIndex = this.event.logIndex.toI32();
-    liquidate.gasPrice = this.event.transaction.gasPrice;
-    liquidate.gasUsed = this.event.receipt ? this.event.receipt!.gasUsed : null;
-    liquidate.gasLimit = this.event.transaction.gasLimit;
-    liquidate.blockNumber = this.event.block.number;
-    liquidate.timestamp = this.event.block.timestamp;
-    liquidate.liquidator = liquidator;
-    liquidate.liquidatee = liquidatee;
-    liquidate.market = this.market.id;
-    liquidate.positions = [liquidatee]; // TODO add position
-    liquidate.asset = asset;
-    liquidate.amount = amount;
-    liquidate.amountUSD = amountUSD;
-    liquidate.profitUSD = profitUSD;
-    liquidate.save();
-
-    this.updateTransactionData(TransactionType.LIQUIDATE, amount, amountUSD);
-    this.updateSnapshotUsage(TransactionType.LIQUIDATEE, liquidatee);
-    this.updateSnapshotUsage(TransactionType.LIQUIDATOR, liquidator);
-
-    return liquidate;
-  }
-
-  createTransfer(
-    asset: Address,
-    sender: Address,
-    receiver: Address,
-    amount: BigInt,
-    amountUSD: BigDecimal
-  ): Transfer {
-    this.account = new AccountClass(
-      sender,
-      this.market,
-      this.protocol,
-      this.event
-    );
-    if (this.account.newAccount()) {
-      this.protocol.cumulativeUniqueUsers += INT_ONE;
-      this.protocol.save();
-    }
-    const recieverAccount = new AccountClass(
-      receiver,
-      this.market,
-      this.protocol,
-      this.event
-    );
-    // receivers are not considered users since they are not spending gas for the transaction
-
-    const transfer = new Transfer(
-      this.event.transaction.hash.concatI32(this.event.logIndex.toI32())
-    );
-    transfer.hash = this.event.transaction.hash;
-    transfer.nonce = this.event.transaction.nonce;
-    transfer.logIndex = this.event.logIndex.toI32();
-    transfer.gasPrice = this.event.transaction.gasPrice;
-    transfer.gasUsed = this.event.receipt ? this.event.receipt!.gasUsed : null;
-    transfer.gasLimit = this.event.transaction.gasLimit;
-    transfer.blockNumber = this.event.block.number;
-    transfer.timestamp = this.event.block.timestamp;
-    transfer.sender = sender;
-    transfer.receiver = receiver;
-    transfer.market = this.market.id;
-    transfer.positions = [sender, receiver]; // TODO add positions
-    transfer.asset = asset;
-    transfer.amount = amount;
-    transfer.amountUSD = amountUSD;
-    transfer.save();
-
-    this.updateTransactionData(TransactionType.TRANSFER, amount, amountUSD);
-    this.updateSnapshotUsage(TransactionType.TRANSFER, sender);
-
-    return transfer;
-  }
-
-  createFlashloan(
-    asset: Address,
-    account: Address,
-    amount: BigInt,
-    amountUSD: BigDecimal
-  ): Flashloan {
-    this.account = new AccountClass(
-      account,
-      this.market,
-      this.protocol,
-      this.event
-    );
-    if (this.account.newAccount()) {
-      this.protocol.cumulativeUniqueUsers += INT_ONE;
-      this.protocol.save();
-    }
-
-    const flashloan = new Flashloan(
-      this.event.transaction.hash.concatI32(this.event.logIndex.toI32())
-    );
-    flashloan.hash = this.event.transaction.hash;
-    flashloan.nonce = this.event.transaction.nonce;
-    flashloan.logIndex = this.event.logIndex.toI32();
-    flashloan.gasPrice = this.event.transaction.gasPrice;
-    flashloan.gasUsed = this.event.receipt ? this.event.receipt!.gasUsed : null;
-    flashloan.gasLimit = this.event.transaction.gasLimit;
-    flashloan.blockNumber = this.event.block.number;
-    flashloan.timestamp = this.event.block.timestamp;
-    flashloan.account = account;
-    flashloan.market = this.market.id;
-    flashloan.asset = asset;
-    flashloan.amount = amount;
-    flashloan.amountUSD = amountUSD;
-    flashloan.save();
-
-    this.updateTransactionData(TransactionType.FLASHLOAN, amount, amountUSD);
-    this.updateSnapshotUsage(TransactionType.FLASHLOAN, account);
-
-    return flashloan;
-  }
-
-  //////////////////
-  //// Updaters ////
-  //////////////////
-
-  private updateSnapshotUsage(transactionType: string, account: Address): void {
+  public updateUsageData(transactionType: string, account: Bytes): void {
     const dailyActiveAccountID = AccountActivity.DAILY.concat("-")
       .concat(account.toHexString())
       .concat("-")
@@ -1105,12 +466,12 @@ export class MarketClass {
 
     this.market.save();
     this.protocol.save();
-    this.usageDailySnapshot.save();
-    this.usageHourlySnapshot.save();
+    this.createOrUpdateUsageDailySnapshot();
+    this.createOrUpdateUsageHourlySnapshot();
     this.marketDailySnapshot.save();
   }
 
-  private updateTransactionData(
+  updateTransactionData(
     transactionType: string,
     amount: BigInt,
     amountUSD: BigDecimal
@@ -1233,113 +594,18 @@ export class MarketClass {
 
     this.protocol.save();
     this.market.save();
-    this.marketDailySnapshot.save();
-    this.marketHourlySnapshot.save();
-    this.financialSnapshot.save();
+    this.createOrUpdateMarketDailySnapshot();
+    this.createOrUpdateMarketHourlySnapshot();
+    this.createOrUpdateFinancials();
     this.usageDailySnapshot.save();
     this.usageHourlySnapshot.save();
-  }
-
-  // used to update tvl, borrow balance, reserves, etc. in market and protocol
-  updateMarketAndProtocolData(
-    inputTokenPriceUSD: BigDecimal,
-    newInputTokenBalance: BigInt,
-    newVariableBorrowBalance: BigInt | null = null,
-    newStableBorrowBalance: BigInt | null = null,
-    newReserveBalance: BigInt | null = null,
-    exchangeRate: BigDecimal | null = null
-  ): void {
-    const mantissaFactorBD = exponentToBigDecimal(
-      this.inputToken.getDecimals()
-    );
-    this.inputToken.updatePrice(inputTokenPriceUSD);
-    this.market.inputTokenPriceUSD = inputTokenPriceUSD;
-    this.market.inputTokenBalance = newInputTokenBalance;
-    if (newVariableBorrowBalance) {
-      this.market.variableBorrowedTokenBalance = newVariableBorrowBalance;
-    }
-    if (newStableBorrowBalance) {
-      this.market.stableBorrowedTokenBalance = newStableBorrowBalance;
-    }
-    if (newReserveBalance) {
-      this.market.reserves = newReserveBalance
-        .toBigDecimal()
-        .div(mantissaFactorBD)
-        .times(inputTokenPriceUSD);
-    }
-    if (exchangeRate) {
-      this.market.exchangeRate = exchangeRate;
-    }
-    const vBorrowAmount = this.market.variableBorrowedTokenBalance
-      ? this.market
-          .variableBorrowedTokenBalance!.toBigDecimal()
-          .div(mantissaFactorBD)
-      : BIGDECIMAL_ZERO;
-    const sBorrowAmount = this.market.stableBorrowedTokenBalance
-      ? this.market
-          .stableBorrowedTokenBalance!.toBigDecimal()
-          .div(mantissaFactorBD)
-      : BIGDECIMAL_ZERO;
-    const totalBorrowed = vBorrowAmount.plus(sBorrowAmount);
-    this.market.totalValueLockedUSD = newInputTokenBalance
-      .toBigDecimal()
-      .div(mantissaFactorBD)
-      .times(inputTokenPriceUSD);
-    this.market.totalBorrowBalanceUSD = totalBorrowed.times(inputTokenPriceUSD);
-    this.market.save();
-
-    let totalValueLockedUSD = BIGDECIMAL_ZERO;
-    let totalBorrowBalanceUSD = BIGDECIMAL_ZERO;
-    const marketList = this.protocol._markets;
-    for (let i = 0; i < marketList.length; i++) {
-      const _market = Market.load(marketList[i]);
-      if (!_market) {
-        log.error("[updateMarketAndProtocolData] Market not found: {}", [
-          marketList[i].toHexString(),
-        ]);
-        continue;
-      }
-      log.error("asdf tvl: {}", [_market.totalValueLockedUSD.toString()]);
-      totalValueLockedUSD = totalValueLockedUSD.plus(
-        _market.totalValueLockedUSD
-      );
-      totalBorrowBalanceUSD = totalBorrowBalanceUSD.plus(
-        _market.totalBorrowBalanceUSD
-      );
-    }
-    this.protocol.totalValueLockedUSD = totalValueLockedUSD;
-    this.protocol.totalBorrowBalanceUSD = totalBorrowBalanceUSD;
-    this.protocol.save();
   }
 
   updateRevenue(
     newTotalRevenueUSD: BigDecimal,
     newProtocolRevenueUSD: BigDecimal,
-    newSupplySideRevenueUSD: BigDecimal,
-    revenueSource: string | null = null
+    newSupplySideRevenueUSD: BigDecimal
   ): void {
-    // update market
-    this.market.cumulativeTotalRevenueUSD =
-      this.market.cumulativeTotalRevenueUSD.plus(newTotalRevenueUSD);
-    this.market.cumulativeProtocolSideRevenueUSD =
-      this.market.cumulativeProtocolSideRevenueUSD.plus(newProtocolRevenueUSD);
-    this.market.cumulativeSupplySideRevenueUSD =
-      this.market.cumulativeSupplySideRevenueUSD.plus(newSupplySideRevenueUSD);
-    this.market.save();
-
-    // update protocol
-    this.protocol.cumulativeTotalRevenueUSD =
-      this.protocol.cumulativeTotalRevenueUSD.plus(newTotalRevenueUSD);
-    this.protocol.cumulativeProtocolSideRevenueUSD =
-      this.protocol.cumulativeProtocolSideRevenueUSD.plus(
-        newProtocolRevenueUSD
-      );
-    this.protocol.cumulativeSupplySideRevenueUSD =
-      this.protocol.cumulativeSupplySideRevenueUSD.plus(
-        newSupplySideRevenueUSD
-      );
-    this.protocol.save();
-
     // update market hourly snapshot
     this.marketHourlySnapshot.hourlyTotalRevenueUSD =
       this.marketHourlySnapshot.hourlyTotalRevenueUSD.plus(newTotalRevenueUSD);
@@ -1351,7 +617,7 @@ export class MarketClass {
       this.marketHourlySnapshot.hourlySupplySideRevenueUSD.plus(
         newSupplySideRevenueUSD
       );
-    this.marketHourlySnapshot.save();
+    this.createOrUpdateMarketHourlySnapshot();
 
     // update market daily snapshot
     this.marketDailySnapshot.dailyTotalRevenueUSD =
@@ -1364,7 +630,7 @@ export class MarketClass {
       this.marketDailySnapshot.dailySupplySideRevenueUSD.plus(
         newSupplySideRevenueUSD
       );
-    this.marketDailySnapshot.save();
+    this.createOrUpdateMarketDailySnapshot();
 
     // update financials snapshot
     this.financialSnapshot.dailyTotalRevenueUSD =
@@ -1377,44 +643,80 @@ export class MarketClass {
       this.financialSnapshot.dailySupplySideRevenueUSD.plus(
         newSupplySideRevenueUSD
       );
-    this.financialSnapshot.save();
-
-    // update RevenueDetails
-    if (revenueSource) {
-      const marketDetails = this.getOrCreateRevenueDetails(this.market.id);
-      let sourceIndex = marketDetails.sources.indexOf(revenueSource);
-      if (sourceIndex != -1) {
-        marketDetails.amountsUSD[sourceIndex] =
-          marketDetails.amountsUSD[sourceIndex].plus(newTotalRevenueUSD);
-        marketDetails.save();
-      }
-      this.market.revenueDetails = marketDetails.id;
-      this.market.save();
-
-      const protocolDetails = this.getOrCreateRevenueDetails(this.protocol.id);
-      sourceIndex = protocolDetails.sources.indexOf(revenueSource);
-      if (sourceIndex != -1) {
-        protocolDetails.amountsUSD[sourceIndex] =
-          protocolDetails.amountsUSD[sourceIndex].plus(newTotalRevenueUSD);
-        protocolDetails.save();
-      }
-      this.protocol.revenueDetails = protocolDetails.id;
-      this.protocol.save();
-    }
+    this.createOrUpdateFinancials();
   }
-}
 
-export class ProtocolData {
-  constructor(
-    public readonly protocolID: Bytes,
-    public readonly protocol: string,
-    public readonly name: string,
-    public readonly slug: string,
-    public readonly network: string,
-    public readonly lendingType: string,
-    public readonly lenderPermissionType: string | null,
-    public readonly borrowerPermissionType: string | null,
-    public readonly collateralizationType: string | null,
-    public readonly riskType: string | null
-  ) {}
+  ///////////////////
+  ///// Helpers /////
+  ///////////////////
+
+  private getSnapshotRates(rates: string[], timeSuffix: string): string[] {
+    const snapshotRates: string[] = [];
+    for (let i = 0; i < rates.length; i++) {
+      const rate = InterestRate.load(rates[i]);
+      if (!rate) {
+        log.warning("[getSnapshotRates] rate {} not found, should not happen", [
+          rates[i],
+        ]);
+        continue;
+      }
+
+      // create new snapshot rate
+      const snapshotRateId = rates[i].concat("-").concat(timeSuffix);
+      const snapshotRate = new InterestRate(snapshotRateId);
+      snapshotRate.rate = rate.rate;
+      if (rate.maturityBlock) snapshotRate.maturityBlock = rate.maturityBlock;
+      snapshotRate.side = rate.side;
+      snapshotRate.type = rate.type;
+      if (rate.tranche) snapshotRate.tranche = rate.tranche;
+      snapshotRate.save();
+      snapshotRates.push(snapshotRateId);
+    }
+    return snapshotRates;
+  }
+
+  private getSnapshotFees(fees: string[], timeSuffix: string): string[] {
+    const snapshotFees: string[] = [];
+    for (let i = 0; i < fees.length; i++) {
+      const rate = InterestRate.load(fees[i]);
+      if (!rate) {
+        log.error("[getSnapshotFees] fee {} not found, should not happen", [
+          fees[i],
+        ]);
+        continue;
+      }
+
+      // create new snapshot rate
+      const snapshotFeeID = fees[i].concat("-").concat(timeSuffix);
+      const snapshotFee = new Fee(snapshotFeeID);
+      snapshotFee.rate = rate.rate;
+      snapshotFee.type = rate.type;
+      snapshotFee.save();
+      snapshotFees.push(snapshotFee.id);
+    }
+    return snapshotFees;
+  }
+
+  private getSnapshotRevenueDetails(
+    currID: Bytes,
+    timeSuffix: i32
+  ): Bytes | null {
+    const currDetails = RevenueDetails.load(currID);
+    if (!currDetails) {
+      log.error(
+        "[getRevenueDetailsSnapshot] Cannot find revenue details id: {}",
+        [currID.toHexString()]
+      );
+      return null;
+    }
+
+    const newDetails = new RevenueDetails(
+      currDetails.id.concat(Bytes.fromI32(timeSuffix))
+    );
+    newDetails.sources = currDetails.sources;
+    newDetails.amountsUSD = currDetails.amountsUSD;
+    newDetails.save();
+
+    return newDetails.id;
+  }
 }
